+++ ddtrace/contrib/redis_utils.py
@@ -1,0 +1,18 @@
+from ddtrace.ext import net
+from ddtrace.ext import redis as redisx
+
+
+def _extract_conn_tags(conn_kwargs):
+    """Transform redis conn info into dogtrace metas"""
+    try:
+        conn_tags = {
+            net.TARGET_HOST: conn_kwargs["host"],
+            net.TARGET_PORT: conn_kwargs["port"],
+            redisx.DB: conn_kwargs.get("db") or 0,
+        }
+        client_name = conn_kwargs.get("client_name")
+        if client_name:
+            conn_tags[redisx.CLIENT_NAME] = client_name
+        return conn_tags
+    except Exception:
+        return {}
